Refactoring ReadUpload
