+++ src/index.ts
@@ -1,17 +1,17 @@
 import './index.css';
-import './modules/results';
+// import './modules/results';
 import faviconUrl from "./assets/vite.svg"
 import txtUrl from "./assets/Ten-pin-results-example.txt"
 import { cloneDeep } from 'lodash-es';
-import { readUpload } from './modules/fileUpload';
-import { getAllPlayerResults } from './modules/calculateResults';
+import { readUpload } from './modules/readUpload';
+import { calculateResults } from './modules/calculateResults';
+import { displayResults } from './modules/displayResults';
 import { deepClone } from '@vitest/utils';
 import { PlayerInt, PlayerRollsInt, HTMLInputEvent } from './types/types';
 
 /*
-
   The app does 4 things:
-    1. Reads a text file that was input.
+    1. Gets entries from a text file.
     2. Calculates results.
     3. Prepares results to be displayed.
     4. Displays results.
@@ -25,61 +25,64 @@ import { PlayerInt, PlayerRollsInt, HTMLInputEvent } from './types/types';
   "Frame" is one of ten scoring rounds constituting a game.
   "Roll" is as single throw of a ball.
   Frames consist of 1,2 or 3 rolls.
-
 */
 
-// DOM Elements:
+// DOM Elements
 const uploadButton = document.querySelector('.upload__button');
 const uploadInput: HTMLInputElement | null = document.querySelector('.upload__input');
 
+
+
+
+
 // Initialization:
 (() => {  
+
   // This listener makes our button upload files:
   uploadButton?.addEventListener('click', (event) => {
     event.preventDefault();
     uploadInput?.click();
-  })
-  
+  });
   // This listener starts the whole read-prepare-display sequence:
   uploadInput?.addEventListener('change', ( event: any ) => {
     event.preventDefault();
-    getPlayerData( event );
-  })
+    getPlayerEntries( event );
+  });
+
 })();
 
-async function getPlayerData( event: HTMLInputEvent ) {
+// Step 1: Get entries from a text file.
+async function getPlayerEntries( event: HTMLInputEvent ) {
+
+  const playerEntries = readUpload( event );
+  processForDisplay( await playerEntries );
+
+};
+
+// Steps 2 & 3: Calculate results, prepare for display and save:
+function processForDisplay ( entries: PlayerRollsInt[] ) {
+
+  const playerEntries: PlayerRollsInt[] = cloneDeep( entries );  
+  const detailedResults = calculateResults( playerEntries );
+  const tableHtml = displayResults( detailedResults );
+  saveToSessionStorage( tableHtml );
   
-  // Here we get the names and rolls of each player:
-  const playerRolls = readUpload( event );
-  console.log('playerRolls', playerRolls)
-  processPlayerEntries( await playerRolls );
+  // Finally, we switch to results.html:
+  if ( tableHtml !== "" ) redirectToResultsPage();
+
 }
 
-// Here we dispatch data to appropriate modules and save data:
-const processPlayerEntries = ( text: PlayerRollsInt[] ) => {
+function saveToSessionStorage( data: string ) {
 
-  const playerEntries: PlayerRollsInt[] = cloneDeep(text);
-  
-  // First, we process text entries into frames:
-  const playerResults = getAllPlayerResults( playerEntries );
-  console.log('player results: ', playerResults);
-  
-
-  // Now we can save player results to sessionstorage
-  saveToSessionStorage( playerResults );
-  
-  // Now we can switch to results.html
-  // if ( playerResults.length !== 0 ) redirectToResultsPage();  
-}
-
-function saveToSessionStorage( data: PlayerInt[] ) {
   const dataCopy = deepClone( data );
   const json = JSON.stringify( dataCopy );
-
   window.sessionStorage.setItem( 'playerResults', json );
+
 }
 
 function redirectToResultsPage() {
+
   const newLocation = `${window.location.origin}/pages/results.html`;
   window.location.replace(newLocation);
+
 }
