+++ src/modules/results.ts
@@ -1,25 +1,13 @@
-import { displayResults } from './displayResults';
+import { FrameInt, PlayerInt } from "../types/types";
 
-
-interface FrameInt {
-  frameId: number;
-  rolls: number[];
-  isStrike: boolean;
-  isSpare: boolean;
-  pointResult: number;
-}
-
-interface PlayerInt {
-  name: string;
-  result: number | null;
-  frames: FrameInt[];
-}
+// DOM Elements
+const tableBody = document.querySelector('.table__body');
 
 console.log('This should work on table page only.');
 
 const getDataFromSessionStorage = () => {
   const data = window.sessionStorage.getItem( 'playerResults' );
-  let playerData: PlayerInt[] = [];
+  let playerData: string = `No results found.`;
 
   if ( data !== null ) {
     playerData = JSON.parse( data );
@@ -28,5 +16,11 @@ const getDataFromSessionStorage = () => {
   return playerData;
 }
 
-const playerData = getDataFromSessionStorage();
-displayResults( playerData );
+// displayResults( playerData );
+
+(function display () {
+  const tableRows = getDataFromSessionStorage();
+  if ( tableBody !== null ) {
+    tableBody.innerHTML = tableRows;
+  };
+})();
